MedicalAi/requirements.txt
Django==5.2.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-cors-headers==4.7.0
djoser==2.3.1
transformers==4.52.4
torch==2.7.0
numpy==2.2.6
Pillow==11.2.1
requests==2.32.3
PyJWT==2.9.0
python-dotenv==1.0.0
huggingface-hub==0.32.3
tokenizers==0.21.1
safetensors==0.5.3